+++ eras/mary/impl/src/Cardano/Ledger/Mary/UTxO.hs
@@ -15,8 +15,9 @@ import Cardano.Ledger.Coin (Coin)
 import Cardano.Ledger.Credential (Credential)
 import Cardano.Ledger.Mary.Core
 import Cardano.Ledger.Mary.Era (MaryEra)
+import Cardano.Ledger.Mary.PolicyID (policyID)
 import Cardano.Ledger.Mary.State ()
-import Cardano.Ledger.Mary.Value (MaryValue (..), policyID)
+import Cardano.Ledger.Mary.Value (MaryValue (..))
 import Cardano.Ledger.Shelley.UTxO (
   ShelleyScriptsNeeded (..),
   getShelleyMinFeeTxUtxo,